Module Item Kebab Menu - Edit
In Module Item list in Modules page the edit form must be shown
and allow a teacher to edit the assignment.

Test Plan:

1. Enable feature flag modules_page_rewrite
2. Select a Course
3. Go to Modules section
4. In modules item list clic the three dots menu
5. Select edit option:
	Verify that form is shown correctly
	Verify that initial values are shown correctly (current item
	title and indent)
6. Change the values in the form: title and indent
7. Click Update button
8. Verify values changed are updated in Module Item List
9. Refresh page and check that item values are correctly persisted.

Note: This behaviour must be the same that is currently working without the
activation of this flag. It is good to first execute the test plan without
the flag and verify that it works in the same way when the flag is active.

flag=modules_page_rewrite
Closes LX-2358

Change-Id: Ie81f77aceab66fba6462bd4f79f6eea27e377f33
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/377046
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>