fix file reference for attachment copies
note: previously it could happen that the
file reference broke for attachments that
have copies if either the original or one
of the copies was deleted. In this commit
we make sure to switch the copies' content
to the placeholder and not just for the
original attachment.

refs: FOO-5167
flag=none

test plan:
- visit "/files" URL and create a folder inside "My File".
- upload an image or a pdf there.
- go to a course and create 2 separate assignment w/ file based submission.
- upload a submission for both in the following way:
-- click "New Attempt" then "Click here to find a file you've already uploaded"
-- chose the file you uploaded in you personal folder.
-- submit assignment.
- go to one of the assignment in SG.
- click on the trash icon to delete the file.
- the original attachment and the copies should be replaced with "file_removed.pdf".
- (previously we got an error message for
the copies)

Change-Id: I73e04ea1488059b426e452133ac93f0658e8b018
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/376649
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Adrián Grúber <adrian.gruber@instructure.com>
Product-Review: Adrián Grúber <adrian.gruber@instructure.com>