Module Header Menu - Send To ..
In Module Header in Modules page the Send To.. form must be shown
and allow a teacher to Send the Module to other user.

Test Plan:


1. Enable feature flag modules_page_rewrite
2. Select a Course
3. Go to Modules section
4. In a module Header clic the three dots menu
5. Select send to option
	Verify form is shown correctly
6. Type an user that is not included in data base. i.e ('Any123')
	Verify 'No results' message is shown
7. Type the name of an existing user:
	Verify that all users whose names matches the value inserted
	are shown in the list and are selectable
8. Select one or more users from selectable list.
	Verify selected users are shown in the input
9. Click one of the selected users shown in input with "x" option
	Verify user is removed from selection list
10. Select at least one user from list and clic "Send" option.
	Verify modal is closed
	Verify item is correctly shared with user

Note: This behavior must be the same that is currently working without the
activation of this flag. It is good to first execute the test plan without
the flag and verify that it works in the same way when the flag is active.

flag=modules_page_rewrite
Closes LX-2366

Change-Id: I0387dae1a70444605da8763df8242b42059a077d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/378136
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>