Upgrade react-redux to allow new hooks usage
flag=none

The one change I am seeing in the release notes is that
it makes some of the HOC components "Pure" when it didn't
used to. I don't see anywhere that this will introduce any
issues with asynchronicity or re-renders that need to
happen but fail to, and tests seem to pass, but this might
take a bit of spot-checking (look for any React components
that import connect from react-redux and make sure they are
still in good shape).

Test plan:
* automated tests pass
* see above if you want to do any spot-checking

Change-Id: I70230d3475998f89bd5dc7b22180347732c41000
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/379263
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>