use Modal in /search/all_courses
part of an update to remove jqueryUI dialog and
replace with InstUI modal

because of the existing styling in the embedded HTML
we need to calculate and set the size of the iframe
to create a modal that fits all the content. Otherwise,
the html footer blocks part of the body.

So to reduce pop-in (caused by embedded html, css, then resizing)
a loading spinner is added while the iframe loads.

flag=none

refs FOO-4693
closes FOO-4849

test-plan:
- in root account settings
	- check "Show a searchabe list..."
- with multiple courses in settings
	- check "Include this course in the public index"
	- check "Let students self-enroll..."
	- check "Add a "Join this Course" link..."
- visit /search/all_courses
- click "Join this Course"
- EXPECT a loading state before the iframe loads
- EXPECT the iframe to load with the self-enrollment view

Change-Id: I719d663e611eba89b1fac6cfde2881a22850ae74
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/380238
QA-Review: Maya Tyner <maya.tyner@instructure.com>
Product-Review: Maya Tyner <maya.tyner@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>