Clarify API documentation for reports_controller arguments
Fixes FOO-5504
flag=none

The API docs for the POST call to start a report is a little
confusing as it starts by documenting an argument called
"parameters" of no type (so it displays in the API doc as
"string" which I guess is the default), but there actually
is not standalone "parameters" argument, instead it is an
array... the API doc does then helpfully give some examples
for things like parameters[course_id] but the "generic"
specification needed to be made clearer. This just specifies
that it is to be an array of "Object" which is the most
generic type that Ruby has.

It's not in the scope of the ticket, but a similar thing
appears in the API docs for Course Reports, so the same
change was made there as well.

Test plan:
* you agree with the change to the YARD docs.

Change-Id: Ic3771833fdf2640eee7be27ad08de1194811fcf6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382308
Reviewed-by: Adrián Grúber <adrian.gruber@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>