Implement drag files into folders
this allows the user to drag 1+ files and or folders currently
in the table into a folder on the table. Only works for files/
folders on the current page. If 1+ rows are selected, then all of
the selected items are moved into the folder

note: ui would ideally have a highlight over the drop target folder
to indicate which folder user is dropping to. However this is an
InstUi issue that will be resolved at a later time.

closes RCX-3480
flag=files_a11y_rewrite

test plan:
- within the table, attempt dragging one file/folder, and
  attempt selecting multiple files/folders and dropping into a folder
- on drag, a file icon with the number of dragged items should appear
- name collisions with the files in the destination folder should be
  handled by the file rename modal.
- attempt dragging a file from outside of the table. This should instead
  trigger the upload sequence and should not show any of the new
  dragging behavior.

Change-Id: I571a9c869f1c2fed75f942988b3565b3084e0d44
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382421
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Kata Virag Keri <kata.keri@instructure.com>