Fix Exception Modal "done" when no exceptions
Previously, when a user clicks the "Done" button in the exception modal
without having added any exceptions, the modal would attempt to create
an empty set of exceptions. This would result in the backend responding
with an error because the exceptions array was empty, causing an error
in the UI.

This commit fixes the exception modal to only submit exceptions if
there are any exceptions to submit. If no exceptions are present, the
modal will simply close without attempting to create an empty set of
exceptions.

fixes INTEROP-9477
flag=lti_registrations_next

Change-Id: Iccc575521f0be74881cfa6e799d20447c559fb3d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/382672
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>