Fix breadcrumb size in right-of-crumb for k5 theme
This commit moves the:
```css
.right-of-crumb {
flex: 1;
}
```
style rule to the k5 theme stylesheet to ensure that the breadcrumb size
takes up as much space as possible in the normal canvas theme.

This style was originally introduced in commit
`11a3fcf3cc494964abb76955c0966cf2ff28095d` to ensure that the course
name in "Canvas for Elementary" mode was centered correctly, but it
was applied to all of Canvas, not only applied to the k5 theme.

test plan:
1. Navigate to a course in Canvas.
2. Ensure that the breadcrumb size is correct and takes up the full
   width of the right-of-crumb area.
3. Navigate to a course and ensure that the right-of-crumb area is
   still rendered correctly.
4. Enable "Canvas for Elementary" mode. by going to Admin -> Settings
-> Checking "Canvas for Elementary" -> Save.
5. Navigate to a page in a course to make sure the course name is still
   centered correctly in the topp nav bar.

fixes: INTEROP-9391
flag=none

Change-Id: I9837f13a9fc2b54e6fc150f78978a29746ea9b25
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/383760
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>